pyproject.toml
[project]
name = "medsupport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "optimum[onnxruntime-gpu]>=1.17.1",
    "polars>=1.32.3",
    "pyturso>=0.4.4",
    "requests>=2.32.5",
    "sentence-transformers[train]>=5.1.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.55.2",
]
